colmines/requirements.txt
django
djangorestframework
django-cassandra-engine
debug_toolbar
